+++ solstice/src/main/java/dev/equo/solstice/Capability.java
@@ -101,40 +101,60 @@ public class Capability implements Comparable<Capability> {
 		return copy;
 	}
 
+	public boolean isSubsetOfElementIn(Iterable<Capability> other) {
+		for (var cap : other) {
+			if (isSubsetOf(cap)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
 	/** Order-insignificant matching to determine if this Capability is a subset of its argument. */
 	public boolean isSubsetOf(Capability other) {
 		return other.isSupersetOf(this);
 	}
 
 	/** Order-insignificant matching to determine if this Capability is a superset of its argument. */
-	public boolean isSupersetOf(Capability other) {
-		if (size() >= other.size()) {
-			return __isSubsetMatch(other, this);
-		} else {
+	public boolean isSupersetOf(Capability shorter) {
+		if (size() < shorter.size()) {
 			return false;
 		}
-	}
-
-	private static boolean __isSubsetMatch(Capability shorter, Capability longer) {
 		for (int i = 0; i < shorter.keyValue.size() / 2; ++i) {
 			var key = shorter.keyValue.get(2 * i);
 			var value = shorter.keyValue.get(2 * i + 1);
 
-			var longerKeyIdx = longer.keyValue.indexOf(key);
+			var longerKeyIdx = keyIdxShortcut(key);
 			if (longerKeyIdx == -1) {
 				// missing key
 				return false;
 			}
-			if (longerKeyIdx % 2 == 1) {
-				throw Unimplemented.onPurpose(
-						"Key has the same content as a value, unlikely to ever happen, straight-forward to fix if it does, please file an issue at https://github.com/equodev/equo-ide");
-			}
-			var longerValue = longer.keyValue.get(longerKeyIdx + 1);
+			var longerValue = keyValue.get(longerKeyIdx + 1);
 			if (!value.equals(longerValue)) {
 				return false;
 			}
 		}
 		return true;
+	}
+
+	public String getValue(String key) {
+		int idx = keyIdxShortcut(key);
+		if (idx == -1) {
+			return null;
+		}
+		return keyValue.get(idx + 1);
+	}
+
+	private int keyIdxShortcut(String key) {
+		int idx = keyValue.indexOf(key);
+		if (idx == -1) {
+			return -1;
+		}
+		if (idx % 2 == 1) {
+			throw Unimplemented.onPurpose(
+					"Key has the same content as a value, unlikely to ever happen, straight-forward to fix if it does, please file an issue at https://github.com/equodev/equo-ide");
+		}
+		return idx;
 	}
 
 	/**
@@ -149,9 +169,9 @@ public class Capability implements Comparable<Capability> {
 							if (lookingForSupersetOf == null) {
 								return a.compareTo(b);
 							} else {
-								if (lookingForSupersetOf == a && b.isSupersetOf(a)) {
+								if (lookingForSupersetOf == a && b.isSupersetOf(lookingForSupersetOf)) {
 									return 0;
-								} else if (lookingForSupersetOf == b && a.isSupersetOf(b)) {
+								} else if (lookingForSupersetOf == b && a.isSupersetOf(lookingForSupersetOf)) {
 									return 0;
 								} else {
 									return a.compareTo(b);