+++ solstice/src/main/java/dev/equo/solstice/NestedJars.java
@@ -19,9 +19,11 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.JarURLConnection;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
+import java.util.Base64;
 import java.util.Comparator;
 import java.util.Enumeration;
 import java.util.List;
@@ -184,14 +186,11 @@ public abstract class NestedJars {
 		try (var toRead = entry.openStream()) {
 			var content = toRead.readAllBytes();
 
-			var md5 = MessageDigest.getInstance("MD5");
-			md5.update(content);
-
 			var jarPath = entry.getPath();
 			var lastSep = Math.max(jarPath.lastIndexOf('!'), jarPath.lastIndexOf('/'));
 			var jarSimpleName = jarPath.substring(lastSep + 1);
 
-			var filename = parentJar + "__" + jarSimpleName + "__" + bytesToHex(md5.digest()) + ".jar";
+			var filename = parentJar + "__" + jarSimpleName + "__" + filenameSafeHash(content) + ".jar";
 			var jarToAdd = new File(nestedJarFolder, filename);
 			if (!jarToAdd.exists() || jarToAdd.length() != content.length) {
 				nestedJarFolder.mkdirs();
@@ -200,20 +199,24 @@ public abstract class NestedJars {
 				}
 			}
 			return Map.entry(entry, jarToAdd);
-		} catch (IOException | NoSuchAlgorithmException e) {
+		} catch (IOException e) {
 			throw Unchecked.wrap(e);
 		}
 	}
 
-	private static final char[] HEX_ARRAY = "0123456789ABCDEF".toCharArray();
+	static String filenameSafeHash(String content) {
+		return filenameSafeHash(content.getBytes(StandardCharsets.UTF_8));
+	}
 
-	private static String bytesToHex(byte[] bytes) {
-		char[] hexChars = new char[bytes.length * 2];
-		for (int j = 0; j < bytes.length; j++) {
-			int v = bytes[j] & 0xFF;
-			hexChars[j * 2] = HEX_ARRAY[v >>> 4];
-			hexChars[j * 2 + 1] = HEX_ARRAY[v & 0x0F];
+	static String filenameSafeHash(byte[] content) {
+		try {
+			MessageDigest md5 = MessageDigest.getInstance("MD5");
+			md5.update(content);
+			String encoded =
+					new String(Base64.getEncoder().encode(md5.digest()), StandardCharsets.US_ASCII);
+			return encoded.replace('/', '-').replace('=', '-');
+		} catch (NoSuchAlgorithmException e) {
+			throw new RuntimeException(e);
 		}
-		return new String(hexChars);
 	}
 }